# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EVE Online PAP Tracking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pap/manual:
    post:
      operationId: createManualPap
      tags: [pap]
      summary: Manually create a PAP adjustment (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManualPapBody"
      responses:
        "201":
          description: Manual PAP record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PapRecord"
...
  /identity-groups:
    get:
      operationId: listIdentityGroups
      tags: [identity]
      responses:
        "200":
          description: Identity groups for the active corporation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdentityGroup"
    post:
      operationId: createIdentityGroup
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityGroupInput"
      responses:
        "201":
          description: Identity group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityGroup"
  /identity-groups/{id}:
    patch:
      operationId: updateIdentityGroup
      tags: [identity]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityGroupInput"
      responses:
        "200":
          description: Identity group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityGroup"
  /identity-groups/{id}/applications:
    post:
      operationId: applyIdentityGroup
      tags: [identity]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                characterId: { type: integer }
                statement: { type: string }
              required: [characterId, statement]
      responses:
        "201":
          description: Application submitted and skill-audited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityApplication"
...
  /identity-applications/{id}:
    patch:
      operationId: reviewIdentityApplication
      tags: [identity]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending_review, needs_information, approved, rejected]
                reviewerNotes: { type: string }
              required: [status]
      responses:
        "200":
          description: Application reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityApplication"
...
components:
  schemas:
...
    PapRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        characterId:
          type: ["integer", "null"]
        fleetId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [fleet, manual, adjustment]
        reason:
          type: ["string", "null"]
        fleetName:
          type: ["string", "null"]
        characterName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - type
        - createdAt
    CreateManualPapBody:
      type: object
      properties:
        userId:
          type: integer
        amount:
          type: number
        reason:
          type: string
      required:
        - userId
        - amount
        - reason
...
    RequiredSkill:
      type: object
      properties:
        skillId: { type: integer }
        name: { type: string }
        level: { type: integer, minimum: 1, maximum: 5 }
      required: [skillId, name, level]
    SkillAuditItem:
      allOf:
        - $ref: "#/components/schemas/RequiredSkill"
        - type: object
          properties:
            trainedLevel: { type: integer }
            passed: { type: boolean }
          required: [trainedLevel, passed]
    SkillAuditResult:
      type: object
      properties:
        checkedAt: { type: string, format: date-time }
        passed: { type: boolean }
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillAuditItem"
      required: [checkedAt, passed, skills]
    IdentityApplication:
      type: object
      properties:
        id: { type: integer }
        corporationId: { type: integer }
        groupId: { type: integer }
        groupName: { type: string }
        groupCategory: { type: string, enum: [combat, management] }
        userId: { type: integer }
        applicantName: { type: ["string", "null"] }
        characterId: { type: integer }
        characterName: { type: string }
        statement: { type: string }
        status:
          type: string
          enum: [pending_skill_audit, pending_review, needs_information, approved, rejected, withdrawn]
        skillAudit:
          oneOf:
            - $ref: "#/components/schemas/SkillAuditResult"
            - type: "null"
        rejectionReason: { type: ["string", "null"] }
        reviewerNotes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, corporationId, groupId, userId, characterId, statement, status, createdAt, updatedAt]
    IdentityGroup:
      type: object
      properties:
        id: { type: integer }
        corporationId: { type: integer }
        name: { type: string }
        category: { type: string, enum: [combat, management] }
        description: { type: string }
        requiredSkills:
          type: array
          items:
            $ref: "#/components/schemas/RequiredSkill"
        permissions:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        isMember: { type: boolean }
        latestApplication:
          oneOf:
            - $ref: "#/components/schemas/IdentityApplication"
            - type: "null"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, corporationId, name, category, description, requiredSkills, permissions, isActive, isMember, createdAt, updatedAt]
    IdentityGroupInput:
      type: object
      properties:
        name: { type: string }
        category: { type: string, enum: [combat, management] }
        description: { type: string }
        requiredSkills:
          type: array
          items:
            $ref: "#/components/schemas/RequiredSkill"
        isActive: { type: boolean }
      required: [name, category, description, requiredSkills]
...